refactor: create check.go, use flagrun and seferio
