+++ go.mod
@@ -11,5 +11,7 @@ require github.com/cubicdaiya/gonp v1.0.4
 
 require (
 	github.com/mackerelio/golib v1.2.2
+	github.com/monitoring-forge/flagrun v0.0.7
+	github.com/monitoring-forge/saferio v0.0.2
 	golang.org/x/sys v0.47.0 // indirect
 )
